compNo and departmanrole checks added

+++ ERP-CustomerSide/src/app/components/Technical/regulatory/regulatory.component.ts
@@ -1,15 +1,15 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { MainService } from '../../service/MainService.service';
+import { MainService } from '../../../service/MainService.service';
 import { CommonModule, DatePipe } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
-import { BlankComponent } from '../blank/blank.component';
-import { SectionComponent } from '../section/section.component';
+import { BlankComponent } from '../../blank/blank.component';
+import { SectionComponent } from '../../section/section.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { SwalService } from '../../service/swal.service';
-import { regulatoryModel } from '../../models/regulatoryModel';
-import { vesselModel } from '../../models/vesselModel';
+import { SwalService } from '../../../service/swal.service';
+import { regulatoryModel } from '../../../models/regulatoryModel';
+import { vesselModel } from '../../../models/vesselModel';
 
 @Component({
   selector: 'app-regulatory-information',
@@ -27,6 +27,7 @@ export class RegulatoryInformationComponent implements OnInit {
   filterExpiredData: boolean = false;
   regulatoryForm: FormGroup;
   selectedRegulatory: regulatoryModel | null = null;
+  canEdit: boolean = false;
 
   constructor(
     private userService: MainService,
@@ -46,6 +47,11 @@ export class RegulatoryInformationComponent implements OnInit {
   }
 
   ngOnInit(): void {
+    const currentUser = this.userService.currentUserValue;
+    const decodedToken = this.userService.decodeToken(currentUser?.accessToken);
+    const departmentId = decodedToken?.Departmant;
+    this.canEdit = departmentId === '2' || departmentId === '3';
+
     this.userService.getAllVessels().subscribe((response: vesselModel[]) => {
       this.vessels = response.map(vessel => vessel.vesselName);
     });
@@ -62,9 +68,8 @@ export class RegulatoryInformationComponent implements OnInit {
       }));
       this.applyFilter();
     }, (error) => {
-      this.swal.callToast('error', 'Can not find any regulatory information',3000,false,'warning');
-    }
-  );
+      this.swal.callToast('error', 'Can not find any regulatory information', 3000, false, 'warning');
+    });
   }
 
   applyFilter() {
@@ -76,18 +81,20 @@ export class RegulatoryInformationComponent implements OnInit {
   }
 
   openModal(content: any, regulatory?: regulatoryModel) {
-    if (regulatory) {
-      this.selectedRegulatory = regulatory;
-      this.regulatoryForm.patchValue({
-        ...regulatory,
-        dueBy: regulatory.dueBy ? this.datePipe.transform(regulatory.dueBy, 'yyyy-MM-dd') : '',
-        implementedDate: regulatory.implementedDate ? this.datePipe.transform(regulatory.implementedDate, 'yyyy-MM-dd') : ''
-      });
-    } else {
-      this.selectedRegulatory = null;
-      this.regulatoryForm.reset();
+    if (this.canEdit) {
+      if (regulatory) {
+        this.selectedRegulatory = regulatory;
+        this.regulatoryForm.patchValue({
+          ...regulatory,
+          dueBy: regulatory.dueBy ? this.datePipe.transform(regulatory.dueBy, 'yyyy-MM-dd') : '',
+          implementedDate: regulatory.implementedDate ? this.datePipe.transform(regulatory.implementedDate, 'yyyy-MM-dd') : ''
+        });
+      } else {
+        this.selectedRegulatory = null;
+        this.regulatoryForm.reset();
+      }
+      this.modalService.open(content, { size: 'lg' });
     }
-    this.modalService.open(content, { size: 'lg' });
   }
 
   closeModal() {
@@ -95,7 +102,7 @@ export class RegulatoryInformationComponent implements OnInit {
   }
 
   onSubmit() {
-    if (this.regulatoryForm.valid) {
+    if (this.canEdit && this.regulatoryForm.valid) {
       if (this.selectedRegulatory) {
         this.updateRegulatory();
       } else {
@@ -139,34 +146,38 @@ export class RegulatoryInformationComponent implements OnInit {
           this.applyFilter();
         }
         this.closeModal();
-        this.swal.callToast('success', 'Regulatory info updated successfully',3000,false,'success');
+        this.swal.callToast('success', 'Regulatory info updated successfully', 3000, false, 'success');
       },
       error: (error) => {
-        this.swal.callToast('error', 'Error updating regulatory info',3000,false,'error');
+        this.swal.callToast('error', 'Error updating regulatory info', 3000, false, 'error');
       }
     });
   }
 
   editInfo(info: regulatoryModel) {
-    this.selectedRegulatory = info;
-    this.regulatoryForm.patchValue({
-      ...info,
-      dueBy: info.dueBy ? this.datePipe.transform(info.dueBy, 'yyyy-MM-dd') : '',
-      implementedDate: info.implementedDate ? this.datePipe.transform(info.implementedDate, 'yyyy-MM-dd') : ''
-    });
-    this.modalService.open(this.modalContent, { size: 'lg' });
+    if (this.canEdit) {
+      this.selectedRegulatory = info;
+      this.regulatoryForm.patchValue({
+        ...info,
+        dueBy: info.dueBy ? this.datePipe.transform(info.dueBy, 'yyyy-MM-dd') : '',
+        implementedDate: info.implementedDate ? this.datePipe.transform(info.implementedDate, 'yyyy-MM-dd') : ''
+      });
+      this.modalService.open(this.modalContent, { size: 'lg' });
+    }
   }
 
   deleteInfo(id: number) {
-    this.userService.deleteRegulatoryInfo(id).subscribe({
-      next: () => {
-        this.regulatoryInfo = this.regulatoryInfo.filter(info => info.id !== id);
-        this.applyFilter();
-        this.swal.callToast('success', 'Regulatory info deleted successfully',3000,false,'success');
-      },
-      error: (error) => {
-        this.swal.callToast('error', 'Error deleting regulatory info',3000,false,'error');
-      }
-    });
+    if (this.canEdit) {
+      this.userService.deleteRegulatoryInfo(id).subscribe({
+        next: () => {
+          this.regulatoryInfo = this.regulatoryInfo.filter(info => info.id !== id);
+          this.applyFilter();
+          this.swal.callToast('success', 'Regulatory info deleted successfully', 3000, false, 'success');
+        },
+        error: (error) => {
+          this.swal.callToast('error', 'Error deleting regulatory info', 3000, false, 'error');
+        }
+      });
+    }
   }
 }
